Separate single/multi processing for ebsd_index
