+++ pyebsdindex/__init__.py
@@ -7,4 +7,18 @@ __credits__ = [
 ]
 __description__ = "Python based tool for Hough/Radon based EBSD indexing"
 __name__ = "pyebsdindex"
-__version__ = "0.1rc2"
+__version__ = "0.1dev0"
+
+
+# Try to import only once
+try:
+    import pyopencl
+    _pyopencl_installed = True
+except ImportError:
+    _pyopencl_installed = False
+
+try:
+    import ray
+    _ray_installed = True
+except ImportError:
+    _ray_installed = False
